[Cache] Finish Redis adapter

+++ src/Symfony/Component/Cache/Adapter/RedisAdapter.php
@@ -11,24 +11,22 @@
 
 namespace Symfony\Component\Cache\Adapter;
 
+use Symfony\Component\Cache\Exception\InvalidArgumentException;
+
 /**
  * @author Aurimas Niekis <aurimas@niekis.lt>
  */
 class RedisAdapter extends AbstractAdapter
 {
-    /**
-     * @var \Redis
-     */
     private $redis;
 
-    /**
-     * @param \Redis $redisConnection
-     * @param string $namespace
-     * @param int    $defaultLifetime
-     */
     public function __construct(\Redis $redisConnection, $namespace = '', $defaultLifetime = 0)
     {
         $this->redis = $redisConnection;
+
+        if (preg_match('#[^-+_.A-Za-z0-9]#', $namespace, $match)) {
+            throw new InvalidArgumentException(sprintf('RedisAdapter namespace contains "%s" but only characters in [-+_.A-Za-z0-9] are allowed.', $match[0]));
+        }
 
         parent::__construct($namespace, $defaultLifetime);
     }
@@ -39,18 +37,13 @@ class RedisAdapter extends AbstractAdapter
     protected function doFetch(array $ids)
     {
         $values = $this->redis->mget($ids);
-
         $index = 0;
         $result = [];
 
         foreach ($ids as $id) {
-            $value = $values[$index++];
-
-            if (false === $value) {
-                continue;
+            if (false !== $value = $values[$index++]) {
+                $result[$id] = unserialize($value);
             }
-
-            $result[$id] = unserialize($value);
         }
 
         return $result;
@@ -67,9 +60,19 @@ class RedisAdapter extends AbstractAdapter
     /**
      * {@inheritdoc}
      */
-    protected function doClear()
+    protected function doClear($namespace)
     {
-        return $this->redis->flushDB();
+        if (!isset($namespace[0])) {
+            $this->redis->flushDB();
+        } else {
+            // As documented in Redis documentation (http://redis.io/commands/keys) using KEYS
+            // can hang your server when it is executed against large databases (millions of items).
+            // Whenever you hit this scale, it is advised to deploy one Redis database per cache pool
+            // instead of using namespaces, so that the above FLUSHDB is used instead.
+            $this->redis->eval(sprintf("local keys=redis.call('KEYS','%s*') for i=1,#keys,5000 do redis.call('DEL',unpack(keys,i,math.min(i+4999,#keys))) end", $namespace));
+        }
+
+        return true;
     }
 
     /**
@@ -87,21 +90,26 @@ class RedisAdapter extends AbstractAdapter
      */
     protected function doSave(array $values, $lifetime)
     {
-        $failed = [];
-        foreach ($values as $key => $value) {
-            $value = serialize($value);
+        $failed = array();
 
-            if ($lifetime < 1) {
-                $response = $this->redis->set($key, $value);
-            } else {
-                $response = $this->redis->setex($key, $lifetime, $value);
-            }
-
-            if (false === $response) {
-                $failed[] = $key;
+        foreach ($values as $id => $v) {
+            try {
+                $values[$id] = serialize($v);
+            } catch (\Exception $e) {
+                $failed[] = $id;
             }
         }
 
-        return count($failed) > 0 ? $failed : true;
+        if (!$this->redis->mSet($values)) {
+            return false;
+        }
+
+        if ($lifetime >= 1) {
+            foreach ($values as $id => $v) {
+                $this->redis->expire($id, $lifetime);
+            }
+        }
+
+        return $failed;
     }
 }
